+++ plugins/metadata/pixiv_ranking/__init__.py
@@ -1,0 +1,3 @@
+from .pixiv_ranking import PixivRankingMetadataProvider
+
+__all__ = ["PixivRankingMetadataProvider"]